backend/requirements.txt
# Core Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# NLP & AI
# spaCy 3.6.x keeps compatibility with scispaCy 0.5.x
spacy==3.6.1
scispacy==0.5.3
transformers>=4.41.0,<6.0.0
torch>=2.2.0
sentence-transformers>=2.7.0
langchain==0.0.350
langchain-community==0.0.2

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.1.0
scikit-image==0.22.0
numpy==1.26.2

# Statistical Analysis
pandas==2.1.3
scipy==1.10.1
scikit-learn==1.3.2

# Groq AI (Free Vision LLM)
groq==0.4.2

# RAG & Vector Store (for semantic food search + AI Coach context)
faiss-cpu==1.7.4

# Utilities
python-dateutil==2.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
email-validator>=2.1.1

# CORS
fastapi-cors==0.0.6

# Visualization & Reporting
matplotlib==3.8.2
seaborn==0.13.0

# Testing
pytest==7.4.4